[KBM] KeyboardManagerCommon refactoring (#10909)
