--- src/modules/keyboardmanager/common/RemapShortcut.cpp
@@ -1,2 +1,0 @@
-#include "pch.h"
-#include "RemapShortcut.h"
